[PE-9324] Remove mobile wallet connect
