+++ packages/mobile/package.json
@@ -96,7 +96,6 @@
     "@stripe/crypto": "0.0.4",
     "@stripe/stripe-js": "1.54.1",
     "@tanstack/react-query": "5.62.7",
-    "@walletconnect/react-native-dapp": "1.8.0",
     "array.prototype.flat": "1.2.5",
     "big-integer": "1.6.51",
     "bn.js": "5.2.0",
