+++ packages/mobile/src/screens/external-wallets/ExternalWalletsModalScreen.tsx
@@ -4,22 +4,22 @@ import { ModalScreen } from 'app/components/core'
 
 import { useAppScreenOptions } from '../app-screen/useAppScreenOptions'
 
-import { WalletConnectScreen } from './WalletConnectScreen'
-import { ConfirmRemoveWalletDrawer, WalletsDrawer } from './components'
+import { ExternalWalletsScreen } from './ExternalWalletsScreen'
+import { ConfirmRemoveWalletDrawer, ConnectNewWalletDrawer } from './components'
 
 const Stack = createNativeStackNavigator()
 
 const screenOptionOverrides = { headerRight: () => null }
 
-export const WalletConnectModalScreen = () => {
+export const ExternalWalletsModalScreen = () => {
   const screenOptions = useAppScreenOptions(screenOptionOverrides)
 
   return (
     <ModalScreen>
       <Stack.Navigator screenOptions={screenOptions}>
-        <Stack.Screen name='Wallets' component={WalletConnectScreen} />
+        <Stack.Screen name='Wallets' component={ExternalWalletsScreen} />
       </Stack.Navigator>
-      <WalletsDrawer />
+      <ConnectNewWalletDrawer />
       <ConfirmRemoveWalletDrawer />
     </ModalScreen>
   )
