+++ packages/mobile/src/screens/external-wallets/components/ConfirmRemoveWalletDrawer.tsx
