+++ packages/mobile/src/screens/external-wallets/components/LinkedWallet.tsx
