+++ packages/mobile/src/screens/external-wallets/components/LinkedWallets.tsx
