+++ packages/mobile/src/screens/external-wallets/components/SolanaPhoneOption.tsx
