+++ packages/mobile/src/screens/external-wallets/components/WalletConnectOption.tsx
