+++ packages/mobile/src/screens/external-wallets/components/index.ts
@@ -1,0 +1,3 @@
+export { LinkedWallets } from './LinkedWallets'
+export { ConnectNewWalletDrawer } from './ConnectNewWalletDrawer'
+export * from './ConfirmRemoveWalletDrawer'
