+++ packages/mobile/src/screens/external-wallets/components/useSolanaPhoneAuthorization.ts
