+++ packages/mobile/src/screens/external-wallets/index.ts
@@ -1,0 +1,1 @@
+export { ExternalWalletsModalScreen } from './ExternalWalletsModalScreen'
