+++ packages/mobile/src/screens/external-wallets/types.ts
