+++ packages/mobile/src/screens/external-wallets/useCanConnectNewWallet.ts
