--- packages/mobile/src/screens/wallet-connect/components/index.ts
@@ -1,6 +1,0 @@
-export { LinkedWallets } from './LinkedWallets'
-export {
-  WalletConnectDrawer as WalletsDrawer,
-  WalletConnectProviderRenderModal
-} from './WalletConnectDrawer'
-export * from './ConfirmRemoveWalletDrawer'
