--- packages/mobile/src/screens/wallet-connect/index.ts
@@ -1,2 +1,0 @@
-export { WalletConnectModalScreen } from './WalletConnectModalScreen'
-export { WalletConnectProvider } from './WalletConnectProvider'
